tools: split histogram2influx into two tools.
This rewrites histogram2influx.py into histogram2csv.py that prodices
CSV files and csv2influx.py that reads the CSV files and pushes the data
into an InfluxDB. This was done to make it easier to work with these
tools, as well as produce output that can be read by other tools (such
as spreadsheet applications) without requiring setting up an InfluxDB
and dashboards. The tools support writing to and reading from standard
output/input, so data can be piped from one to the other to avoid
hitting the disk.