frame: factor-out byte stream generation, add properties
This factors out generating the byte stream for sending from class
SendFrame and moves it to a dedicated function "make_frame". This
reduces complexity and memory consumption for cases where the values
that were used to craft the frame are of no interest after sending, so
one can now simply store the byte stream without having to instantiate a
class each time to generate it.