tools: overhaul time series collector
Complete overhaul of the time series collector script. The script was
renamed from "histogram2csv.py" to "timeseries2csv.py", the old name was
plain wrong. The script now deals with all the time series that are
available from the device, which introduces a lot of extra complexity.

The addition of week, month and year series neccessitated adding a
--count parameter to allow specifying how far back to go in time. This
may lead to the script not terminating if the requested range goes past
the devices buffer. This has always been a problem (and is handled for
"year" resolution only because the device inserts a sentinel value
dating to the year 2000).