cli: support simple time series query
Adds support to the CLI for querying for time series data. When a known
time series OID is specified (detected by looking at the response data
type), the current timestamp sent as starting point and whatever the
device sends is returned to the user. Note that the device may sent
varying amounts of data with each request, so care must be taken when
parsing the data. If support for more specific time ranges is required,
the "timeseries2csv.py" tool in the tools subdirectory is the way to go.