Rework event table
The event entries are now storing their payload in bytes. With the
current knowledge, it makes little sense to interpret a select few
entries. This commit makes all the fields that are varying byte
sequences, making it easier to interpret the values.
This also adds simple event table support to the CLI.