Fix completion-related breakage
Click introduced a new API in 8.1 and made it backwards-incompatible.
They also changed the invocation with 8.0.

Documentation was updated to include the new way to get the completion,
along with some examples for Linux distributions that support the major
version 7 or 8 of click.

To support most click versions, the completion was changed to work from
click 7.0 onwards, by conditionally changing the argument handed to the
"option" decorator. Additionally, the completion function now simply
uses kwargs, support the different calling conventions throughout the
versions.

Since it isn't yet known if there will be more breakage in future
versions, this commit limits the allowed / supported versions of click
to less than 8.2, which is expected to be the next release.

Thanks to @poggenpower for finding this.

Fixes #17
Fixes #18